Move get_fill_canvas into VShaderWrapper
